feat: Recruitment 값 타입 구조 변경
