+++ src/main/resources/application.yml
@@ -1,1 +1,4 @@
-
+spring:
+  datasource:
+    username: sa
+    driver-class-name: org.h2.Driver
