+++ src/test/java/com/clova/anifriends/base/BaseControllerTest.java
@@ -3,11 +3,13 @@ package com.clova.anifriends.base;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 
 import com.clova.anifriends.base.config.RestDocsConfig;
+import com.clova.anifriends.domain.recruitment.service.RecruitmentService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
 import org.springframework.restdocs.RestDocumentationContextProvider;
 import org.springframework.restdocs.RestDocumentationExtension;
@@ -30,6 +32,9 @@ public abstract class BaseControllerTest {
 
     @Autowired
     protected RestDocumentationResultHandler restDocs;
+
+    @MockBean
+    protected RecruitmentService recruitmentService;
 
     @BeforeEach
     void setUp(
