+++ src/test/java/com/clova/anifriends/domain/recruitment/wrapper/RecruitmentContentTest.java
