+++ src/test/java/com/clova/anifriends/domain/recruitment/wrapper/RecruitmentTitleTest.java
