+++ src/test/java/com/clova/anifriends/global/exception/GlobalExceptionHandlerTest.java
